LaTeX reader: Allow any block in table cells
In particular this allows us to parse nested tables. Previously a cell would be
delimited by the next ampersand. This makes it impossible to parse a table
containing an ampersand within a table cell.

This commit adds an EscapeMode to the LaTeX parser state that tells us whether
we explicitly want to escape special TeX characters (e.g. when the input is
markdown). Otherwise, we are treating the input as LaTeX code, so we should
not parse them as strings in the parser for inline.